requirements_dev.txt
alabaster==0.7.10
anyconfig
argh==0.26.2
arrow==0.12.1
attrs==18.1.0
binaryornot==0.4.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
future==0.16.0
github3.py==1.1.0
idna==2.7
Jinja2==2.10
jinja2-time==0.2.0
pathlib==1.0.1
pathtools==0.1.2
pluggy==0.6.0
poyo==0.4.1
py==1.5.4
PyJWT==1.6.4
pytest==3.4.2
pytest-cookies==0.3.0
python-dateutil==2.7.3
pywildcard==1.0.8
PyYAML==3.13
requests==2.19.1
six==1.11.0
tox==2.9.1
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
watchdog==0.8.3
whichcraft==0.4.1
yml==0.0.1
configobj==5.0.6
./plugins/package/plugins-1.1.tar.gz
